Let BarCodeScannerReader handle lack of camera permission
- Added the state "barCodeScanned" to prevent a bug with the BarCodeScanner component
where it will keep scanning even if you have navigated away and are on a new page. This fix is not ideal: if
you were to go back to the scanner page after you had scanned, it would not scan again. However, if you return
to the homepage and scan again it will work. This is better than it continuously scanning and messing up the routing.